fix: Paging Admin Dashboard

+++ identity-service/src/main/java/com/sketchnotes/identityservice/service/BlogServiceImpl.java
@@ -74,7 +74,7 @@ public class BlogServiceImpl implements BlogService {
                 responses,
                 blogs.getNumber(),
                 blogs.getSize(),
-                (int) blogs.getTotalElements(),
+                blogs.getTotalElements(),
                 blogs.getTotalPages(),
                 blogs.isLast()
         );
