+++ identity-service/src/main/java/com/sketchnotes/identityservice/service/UserService.java
@@ -68,7 +68,7 @@ public class UserService implements IUserService {
                 userResponses,
                 users.getNumber(),
                 users.getSize(),
-                (int) users.getTotalElements(),
+                users.getTotalElements(),
                 users.getTotalPages(),
                 users.isLast()
         );
